move/rename experiment/nursery to greenhouse
