+++ BUILD.bazel
@@ -37,6 +37,7 @@ filegroup(
         "//gcsweb/cmd/gcsweb:all-srcs",
         "//gcsweb/pkg/version:all-srcs",
         "//ghclient:all-srcs",
+        "//greenhouse:all-srcs",
         "//hack:all-srcs",
         "//images/bootstrap/barnacle:all-srcs",
         "//jenkins:all-srcs",
