+++ greenhouse/diskutil/BUILD.bazel
@@ -1,12 +1,12 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["cache.go"],
-    importpath = "k8s.io/test-infra/experiment/nursery/diskcache",
+    srcs = ["diskutil.go"],
+    importpath = "k8s.io/test-infra/greenhouse/diskutil",
     visibility = ["//visibility:public"],
     deps = [
-        "//experiment/nursery/diskutil:go_default_library",
+        "//vendor/github.com/djherbis/atime:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
     ],
 )
@@ -23,11 +23,4 @@ filegroup(
     srcs = [":package-srcs"],
     tags = ["automanaged"],
     visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "go_default_test",
-    srcs = ["cache_test.go"],
-    embed = [":go_default_library"],
-    deps = ["//vendor/k8s.io/apimachinery/pkg/util/sets:go_default_library"],
 )
